+++ packages/frontend/editor-ui/src/features/workflows/workflowDiff/DiffBadge.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { NodeDiffStatus } from '@/features/workflows/workflowDiff/useWorkflowDiff';
+import { NodeDiffStatus } from 'n8n-workflow';
 import { computed } from 'vue';
 const props = defineProps<{
 	type: NodeDiffStatus;
