+++ packages/workflow/src/index.ts
@@ -74,6 +74,7 @@ export * from './node-parameters/filter-parameter';
 export * from './node-parameters/parameter-type-validation';
 export * from './node-parameters/path-utils';
 export * from './evaluation-helpers';
+export * from './workflow-diff';
 
 export type {
 	DocMetadata,
