+++ packages/workflow/src/workflow-diff.ts
@@ -1,0 +1,64 @@
+import isEqual from 'lodash-es/isEqual';
+import pick from 'lodash-es/pick';
+
+export const enum NodeDiffStatus {
+	Eq = 'equal',
+	Modified = 'modified',
+	Added = 'added',
+	Deleted = 'deleted',
+}
+
+export type NodeDiff<T> = {
+	status: NodeDiffStatus;
+	node: T;
+};
+
+export type WorkflowDiff<T> = Map<string, NodeDiff<T>>;
+
+export function compareNodes<T extends { id: string }>(
+	base: T | undefined,
+	target: T | undefined,
+): boolean {
+	const propsToCompare = ['name', 'type', 'typeVersion', 'webhookId', 'credentials', 'parameters'];
+
+	const baseNode = pick(base, propsToCompare);
+	const targetNode = pick(target, propsToCompare);
+
+	return isEqual(baseNode, targetNode);
+}
+
+export function compareWorkflowsNodes<T extends { id: string }>(
+	base: T[],
+	target: T[],
+	nodesEqual: (base: T | undefined, target: T | undefined) => boolean = compareNodes,
+): WorkflowDiff<T> {
+	const baseNodes = base.reduce<Map<string, T>>((acc, node) => {
+		acc.set(node.id, node);
+		return acc;
+	}, new Map());
+
+	const targetNodes = target.reduce<Map<string, T>>((acc, node) => {
+		acc.set(node.id, node);
+		return acc;
+	}, new Map());
+
+	const diff: WorkflowDiff<T> = new Map();
+
+	for (const [id, node] of baseNodes.entries()) {
+		if (!targetNodes.has(id)) {
+			diff.set(id, { status: NodeDiffStatus.Deleted, node });
+		} else if (!nodesEqual(baseNodes.get(id), targetNodes.get(id))) {
+			diff.set(id, { status: NodeDiffStatus.Modified, node });
+		} else {
+			diff.set(id, { status: NodeDiffStatus.Eq, node });
+		}
+	}
+
+	for (const [id, node] of targetNodes.entries()) {
+		if (!baseNodes.has(id)) {
+			diff.set(id, { status: NodeDiffStatus.Added, node });
+		}
+	}
+
+	return diff;
+}
